implementing payment methods

+++ src/main/java/paymentMethods/PaymentMethod.java
@@ -5,5 +5,5 @@ public interface PaymentMethod {
      *
      * @return the transaction id
      */
-    String pay(float amount);
+    String pay(double amount);
 }
